checkpint before boot everything together

+++ storage/buckets.py
@@ -20,8 +20,21 @@ def create_bucket(name, zone='eu-west-1'):
         print(e)
 
 
+def get_bucket_by_name(name):
+    client = boto3.client("s3")
+    response = client.list_buckets()
+
+    for itm in response["Buckets"]:
+        if itm["Name"] == name:
+            return itm
+
+
 def delete_bucket(name):
     client = boto3.client("s3")
+    bucket = get_bucket_by_name(name)
+
+    if not bucket:
+        return
 
     try:
         response = client.delete_bucket(
